requirements.txt
django==4.2
djangorestframework==3.14.0
pillow==10